requirements.txt
# Core dependencies
pygame==2.5.2

# Raspberry Pi specific (only needed on Pi)
RPi.GPIO==0.7.1; platform_machine == "aarch64" or platform_machine == "armv7l"
